Add files via upload
InvoiceCreator is a Python application that generates invoices for small businesses. It is built using Python 3 and includes a simple user interface that allows you to enter product details and generate an invoice with just a few clicks. The application includes features such as the ability to add new products, edit existing products, and generate reports on sales and revenue. InvoiceCreator is designed to be easy to use and customizable, allowing you to create professional-looking invoices in minutes.